Move make computer guess to own file
