+++ src/App.tsx
@@ -5,6 +5,7 @@ import './index.scss';
 import { CellStates, PositionArray, ShipInfo } from './types';
 import Board from './components/Board';
 import HeatMapBoard from './components/HeatMapBoard';
+import { useMakeComputerGuess } from './logic/makeComputerGuess';
 
 import { placeShips } from './logic/placeShips';
 import { difficultyClass, ai } from './ai-behaviour';
@@ -20,42 +21,8 @@ export const shipTypes: ShipInfo[] = [
 ];
 
 const GameBoards = () => {
-  const { userShips, computerShips, setUserShips, playerTurn, setPlayerTurn } = useContext(GameContext);
-
-  const makeComputerGuess = React.useCallback(() => {
-    const heatMap = calculateHeatMap(userShips);
-    const flatHeatMap = heatMap.flat();
-
-    // Find the highest value that isn't ai.heatMapSimulations
-    let maxValue = -1;
-    let maxValueIndex = -1;
-
-    flatHeatMap.forEach((value, index) => {
-      if (value !== ai.heatMapSimulations && value > maxValue) {
-        maxValue = value;
-        maxValueIndex = index;
-      }
-    });
-
-    const y = Math.floor(maxValueIndex / 10);
-    const x = maxValueIndex % 10;
-
-    console.log('Computer making guess', 'y', y, 'x', x);
-
-    // There's a ship here and we haven't guessed it yet
-    if (userShips[y][x]?.status === CellStates.unguessed) {
-      const newUserShips = [...userShips];
-      newUserShips[y][x] = { name: userShips[y][x]?.name || null, status: CellStates.hit };
-      setUserShips(newUserShips);
-    }
-
-    // If there's no ship here, we've missed
-    if (!userShips[y][x]) {
-      const newUserShips = [...userShips];
-      newUserShips[y][x] = { name: null, status: CellStates.miss };
-      setUserShips(newUserShips);
-    }
-  }, [userShips, setUserShips]);
+  const { userShips, computerShips, playerTurn, setPlayerTurn } = useContext(GameContext);
+  const makeComputerGuess = useMakeComputerGuess();
 
   useEffect(() => {
     if (playerTurn === 'computer') {
