+++ src/components/UserGuessBoard.test.tsx
@@ -274,6 +274,7 @@ describe('UserGuessBoard', () => {
             userShips: mockComputerShips,
             setUserShips: () => {},
             setComputerShips: setComputerShips,
+            setPlayerTurn: () => {},
           } as any
         }
       >
